refactor: addLog 파일 분리
